+++ app/Core/Models/Permission.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Core\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Permission extends Model
+{
+    //
+}