+++ app/Core/Policies/TaskPolicy.php
@@ -4,6 +4,7 @@ namespace App\Core\Policies;
 
 use App\Core\Models\Task;
 use App\Core\Models\User;
+use App\Authorization\Authorization;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class TaskPolicy
@@ -18,7 +19,7 @@ class TaskPolicy
      */
     public function create(User $user)
     {
-        return $user->hasPermissionTo('create task.' . request()->input('taskable_type') . '->' . request()->input('taskable_id'));
+        return (new Authorization($user))->userHasPermissionTo('create', 'task', request('group_type'), request('group_id'));
     }
 
     /**
@@ -30,7 +31,7 @@ class TaskPolicy
      */
     public function update(User $user, Task $task)
     {
-        return $user->hasPermissionTo('edit task.' . $task->taskable_type . '->' . $task->taskable->id);
+        return (new Authorization($user))->userHasPermissionTo('update', 'task', $task->id, true, request('group_type'), request('group_id'));
     }
 
     /**
@@ -42,6 +43,6 @@ class TaskPolicy
      */
     public function delete(User $user, Task $task)
     {
-        return $user->hasPermissionTo('delete task.' . $task->taskable_type . '->' . $task->taskable->id);
+        return (new Authorization($user))->userHasPermissionTo('delete', 'task', $task->id, true, request('group_type'), request('group_id'));
     }
 }