+++ database/seeds/HeadquarterOfficeSeeder.php
@@ -1,10 +1,9 @@
 <?php
 
 use Carbon\Carbon;
-use App\Models\User;
-use App\Models\Office;
+use App\Core\Models\User;
+use App\Core\Models\Office;
 use Illuminate\Database\Seeder;
-use Spatie\Permission\Models\Permission;
 
 class HeadquarterOfficeSeeder extends Seeder
 {
@@ -25,11 +24,6 @@ class HeadquarterOfficeSeeder extends Seeder
         ]);
         $office = Office::where('name', 'Headquarter')->first();
         $user->offices()->attach($office->id);
-        $user->givePermissionTo([
-            Permission::create(['name' => 'view office->' . $office->id]),
-            Permission::create(['name' => 'create discussion.office->' . $office->id]),
-            Permission::create(['name' => 'create task.office->' . $office->id]),
-        ]);
         $office->huntable()->create([
             'name'        => $office->name,
             'description' => $office->description,