+++ resources/assets/js/components/partials/profileDropdown.vue
@@ -1,13 +1,13 @@
 <template>
 <div class="px-4 border-l flex items-center cursor-pointer p-4">
-  <div id="profile-dropdown" class="flex flex-row items-center" @click="toggleProfileDropdown">
+  <div id="profile-dropdown" class="flex flex-row items-center" @click="toggleProfileDropdown" v-click-outside="hideProfileDropdown">
     <img class="w-10 h-10 rounded-full md:mr-2" :src="generateUrl(user.avatar)">
     <span class="text-grey-darker text-base no-underline hidden md:block">
       {{ user.name }}
       <font-awesome-icon :icon="faAngleDown"></font-awesome-icon>
     </span>
   </div>
-  <div v-if="profileDropdownShown" id="profile-menu" class="absolute bg-white w-48 pin-r mr-2 py-1 shadow-lg rounded z-50" style="top:3.5rem;" v-click-outside="hideProfileDropdown">
+  <div v-if="profileDropdownShown" id="profile-menu" class="absolute bg-white w-48 pin-r mr-2 py-1 shadow-lg rounded z-50" style="top:3.5rem;">
     <a class="px-4 py-2 hover:bg-teal hover:text-white no-underline text-grey-dark block font-medium" :href="profileUrl">
       <span class="w-6 inline-block">
         <font-awesome-icon :icon="faUser" class="pr-1"></font-awesome-icon>