+++ tests/Feature/StatusTest.php
@@ -9,12 +9,12 @@ class StatusTest extends TestCase
     public function setUp()
     {
         parent::setUp();
-        $this->status = factory(\App\Models\Status::class)->create();
+        $this->status = factory(\App\Core\Models\Status::class)->create();
     }
 
     public function test_create_new_status()
     {
-        $status = factory(\App\Models\Status::class)->make();
+        $status = factory(\App\Core\Models\Status::class)->make();
 
         $this->actingAs($this->user)->post('/statuses', [
             'name'      => $status->name,